Add stripe payment UI

+++ Client/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, OnInit, Input, AfterViewInit, ViewChild, OnDestroy } from '@angular/core';
 import { FormGroup } from '@angular/forms';
 import { BasketService } from 'src/app/basket/basket.service';
 import { CheckoutService } from '../checkout.service';
@@ -6,14 +6,28 @@ import { ToastrService } from 'ngx-toastr';
 import { Basket } from 'src/app/shared/_models/basket';
 import { OrderToCreate } from 'src/app/shared/_models/order';
 import { Router, NavigationExtras } from '@angular/router';
+import { environment } from 'src/environments/environment';
+
+declare var Stripe;
 
 @Component({
   selector: 'app-checkout-payment',
   templateUrl: './checkout-payment.component.html',
   styleUrls: ['./checkout-payment.component.scss'],
 })
-export class CheckoutPaymentComponent implements OnInit {
+export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
   @Input() checkoutForm: FormGroup;
+  @ViewChild('cardNumber', { static: true }) cardNumberElement;
+  @ViewChild('cardExpiry', { static: true }) cardExpiryElement;
+  @ViewChild('cardCvc', { static: true }) cardCvcElement;
+
+  stripe: any;
+  cardNumber: any;
+  cardExpiry: any;
+  cardCvc: any;
+  cardErrors: any;
+  cardHandler = this.onChange.bind(this);
+  loading = false;
 
   constructor(
     private basketService: BasketService,
@@ -22,21 +36,67 @@ export class CheckoutPaymentComponent implements OnInit {
     private router: Router
   ) {}
 
-  ngOnInit(): void {}
+  ngAfterViewInit(): void {
+    this.stripe = Stripe(environment.stripeKey);
+    const elements = this.stripe.elements();
 
-  submitOrder() {
+    this.cardNumber = elements.create('cardNumber');
+    this.cardNumber.mount(this.cardNumberElement.nativeElement);
+    this.cardNumber.addEventListener('change', this.cardHandler);
+
+    this.cardExpiry = elements.create('cardExpiry');
+    this.cardExpiry.mount(this.cardExpiryElement.nativeElement);
+    this.cardExpiry.addEventListener('change', this.cardHandler);
+
+    this.cardCvc = elements.create('cardCvc');
+    this.cardCvc.mount(this.cardCvcElement.nativeElement);
+    this.cardCvc.addEventListener('change', this.cardHandler);
+  }
+
+  ngOnDestroy() {
+    this.cardNumber.destroy();
+    this.cardCvc.destroy();
+    this.cardExpiry.destroy();
+  }
+
+  onChange({ error }) {
+    this.cardErrors = error.message;
+  }
+
+  async submitOrder() {
+    this.loading = true;
+
     const basket = this.basketService.getCurrentBasketValue();
-    const orderToCreate = this.getOrderToCreate(basket);
+    const createdOrder = await this.createOrder(basket);
+    const paymentResult = await this.confirmPaymentWithStripe(basket);
 
-    this.checkoutService.createOrder(orderToCreate).subscribe(
-      (order) => {
-        this.toastr.success('Order created successfully');
+    try {
+      if (paymentResult.paymentIntent) {
         this.basketService.deleteLocalBasket(basket.id);
-        const navigationExtras: NavigationExtras = { state: order };
+        const navigationExtras: NavigationExtras = { state: createdOrder };
         this.router.navigate(['checkout/success'], navigationExtras);
+      } else {
+        this.toastr.error(paymentResult.error.message);
+      }
+    } catch (error) {
+      console.log(error);
+    }
+
+    this.loading = false;
+  }
+
+  private async confirmPaymentWithStripe(basket: Basket) {
+    return this.stripe.confirmCardPayment(basket.clientSecret, {
+      payment_method: {
+        card: this.cardNumber,
+        billing_details: { name: this.checkoutForm.get('paymentForm').get('nameOnCard').value },
       },
-      (err) => this.toastr.error(err.message)
-    );
+    });
+  }
+
+  private async createOrder(basket: Basket) {
+    const orderToCreate = this.getOrderToCreate(basket);
+    return this.checkoutService.createOrder(orderToCreate).toPromise();
   }
 
   private getOrderToCreate(basket: Basket): OrderToCreate {
